pyproject.toml
[project]
name = "ai-girlfriend-agent"
version = "0.1.0"
description = "基于微信的AI女友聊天机器人"
authors = [{name = "AI Girlfriend Agent Team"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "itchat-uos>=1.5.0",
    "openai>=1.6.0",
    "dashscope>=1.14.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "redis>=5.0.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.0",
    "schedule>=1.2.0",
    "pytz>=2023.3",
    "python-dateutil>=2.8.0",
    "jieba>=0.42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
prod = [
    "gunicorn>=21.0.0",
]

[project.scripts]
aigf = "src.main:main"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | data
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip = [".git", ".venv", "data"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]
exclude = [".git", ".venv", "data"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]
